+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New table output layout SolidTable.
 - Allow to use table layout and separator options for other commands.
 - Add raw formatter "raw_fmt".
+- Add "skip-if-no-columns" option.
 
 ### Fixed